requirements.txt
langchain>=1.0.2
langchain-google-genai>=4.0.0
python-dotenv>=1.1.1
